+++ src/vivarium/framework/results/interface.py
@@ -14,6 +14,7 @@ from typing import TYPE_CHECKING, Any, NamedTuple, Sequence, Union
 import pandas as pd
 from pandas.core.groupby.generic import DataFrameGroupBy
 
+from vivarium.component import DEFAULT_EVENT_PRIORITY
 from vivarium.framework.event import Event
 from vivarium.framework.lifecycle import lifecycle_states
 from vivarium.framework.results.observation import (
@@ -195,6 +196,7 @@ class ResultsInterface(Interface):
         pop_filter: str = "",
         include_untracked: bool = False,
         when: str = lifecycle_states.COLLECT_METRICS,
+        priority: int = DEFAULT_EVENT_PRIORITY,
         requires_attributes: list[str] = [],
         results_updater: ResultsUpdater = _required_function_placeholder,
         results_formatter: ResultsFormatter = _default_stratified_observation_formatter,
@@ -219,6 +221,8 @@ class ResultsInterface(Interface):
         when
             Name of the lifecycle phase the observation should happen. Valid values are:
             "time_step__prepare", "time_step", "time_step__cleanup", or "collect_metrics".
+        priority
+            The priority level of the lifecycle phase (see `when`) that this observation will record.
         requires_attributes
             The population attributes that are required by the `aggregator`.
         results_updater
@@ -249,6 +253,7 @@ class ResultsInterface(Interface):
             name=name,
             population_filter=PopulationFilter(pop_filter, include_untracked),
             when=when,
+            priority=priority,
             requires_attributes=requires_attributes,
             results_updater=results_updater,
             results_formatter=results_formatter,
@@ -265,6 +270,7 @@ class ResultsInterface(Interface):
         pop_filter: str = "",
         include_untracked: bool = False,
         when: str = lifecycle_states.COLLECT_METRICS,
+        priority: int = DEFAULT_EVENT_PRIORITY,
         requires_attributes: list[str] = [],
         results_gatherer: ResultsGatherer = _required_function_placeholder,
         results_updater: ResultsUpdater = _required_function_placeholder,
@@ -286,6 +292,8 @@ class ResultsInterface(Interface):
         when
             Name of the lifecycle phase the observation should happen. Valid values are:
             "time_step__prepare", "time_step", "time_step__cleanup", or "collect_metrics".
+        priority
+            The priority level of the lifecycle phase (see `when`) that this observation will record.
         requires_attributes
             The population attributes that are required by the `results_gatherer`.
         results_gatherer
@@ -312,6 +320,7 @@ class ResultsInterface(Interface):
             name=name,
             population_filter=PopulationFilter(pop_filter, include_untracked),
             when=when,
+            priority=priority,
             requires_attributes=requires_attributes,
             results_updater=results_updater,
             results_gatherer=results_gatherer,
@@ -325,6 +334,7 @@ class ResultsInterface(Interface):
         pop_filter: str = "",
         include_untracked: bool = False,
         when: str = lifecycle_states.COLLECT_METRICS,
+        priority: int = DEFAULT_EVENT_PRIORITY,
         requires_attributes: list[str] = [],
         results_formatter: ResultsFormatter = _default_stratified_observation_formatter,
         additional_stratifications: list[str] = [],
@@ -355,6 +365,8 @@ class ResultsInterface(Interface):
         when
             Name of the lifecycle phase the observation should happen. Valid values are:
             "time_step__prepare", "time_step", "time_step__cleanup", or "collect_metrics".
+        priority
+            The priority level of the lifecycle phase (see `when`) that this observation will record.
         requires_attributes
             The population attributes that are required by the `aggregator`.
         results_formatter
@@ -377,6 +389,7 @@ class ResultsInterface(Interface):
             name=name,
             population_filter=PopulationFilter(pop_filter, include_untracked),
             when=when,
+            priority=priority,
             requires_attributes=requires_attributes,
             results_formatter=results_formatter,
             additional_stratifications=additional_stratifications,
@@ -392,6 +405,7 @@ class ResultsInterface(Interface):
         pop_filter: str = "",
         include_untracked: bool = False,
         when: str = lifecycle_states.COLLECT_METRICS,
+        priority: int = DEFAULT_EVENT_PRIORITY,
         requires_attributes: list[str] = [],
         results_formatter: ResultsFormatter = _default_unstratified_observation_formatter,
         to_observe: Callable[[Event], bool] = lambda event: True,
@@ -418,6 +432,8 @@ class ResultsInterface(Interface):
         when
             Name of the lifecycle phase the observation should happen. Valid values are:
             "time_step__prepare", "time_step", "time_step__cleanup", or "collect_metrics".
+        priority
+            The priority level of the lifecycle phase (see `when`) that this observation will record.
         requires_attributes
             The population attributes that are required by the `aggregator`.
         results_formatter
@@ -430,6 +446,7 @@ class ResultsInterface(Interface):
             name=name,
             population_filter=PopulationFilter(pop_filter, include_untracked),
             when=when,
+            priority=priority,
             requires_attributes=requires_attributes,
             results_formatter=results_formatter,
             to_observe=to_observe,