+++ src/vivarium/framework/results/observation.py
@@ -30,7 +30,9 @@ import pandas as pd
 from pandas.api.types import CategoricalDtype
 from pandas.core.groupby.generic import DataFrameGroupBy
 
+from vivarium.component import DEFAULT_EVENT_PRIORITY
 from vivarium.exceptions import VivariumError
+from vivarium.framework import lifecycle
 from vivarium.framework.event import Event
 from vivarium.framework.results.stratification import Stratification, get_original_col_name
 
@@ -84,6 +86,8 @@ class Observation(ABC):
     """Method or function that determines whether to perform an observation on this Event."""
     stratifications: tuple[Stratification, ...] | None = None
     """Optional tuple of the Stratifications this observation should use."""
+    priority: int = DEFAULT_EVENT_PRIORITY
+    """The priority level of the lifecycle phase (see `when`) that this observation will record."""
 
     def observe(
         self,
@@ -140,6 +144,8 @@ class UnstratifiedObservation(Observation):
         Method or function that formats the raw observation results.
     to_observe
         Method or function that determines whether to perform an observation on this Event.
+    priority
+        The priority level of the lifecycle phase (see `when`) that this observation will record.
 
     """
 
@@ -153,6 +159,7 @@ class UnstratifiedObservation(Observation):
         results_updater: Callable[[pd.DataFrame, pd.DataFrame], pd.DataFrame],
         results_formatter: Callable[[str, pd.DataFrame], pd.DataFrame],
         to_observe: Callable[[Event], bool] = lambda event: True,
+        priority: int = DEFAULT_EVENT_PRIORITY,
     ):
         def _wrap_results_gatherer(
             df: pd.DataFrame | DataFrameGroupBy[tuple[str, ...] | str, bool],
@@ -175,6 +182,7 @@ class UnstratifiedObservation(Observation):
             results_updater=results_updater,
             results_formatter=results_formatter,
             to_observe=to_observe,
+            priority=priority,
         )
 
     @classmethod
@@ -224,6 +232,8 @@ class StratifiedObservation(Observation):
         Method or function that computes the quantity for this observation.
     to_observe
         Method or function that determines whether to perform an observation on this Event.
+    priority
+        The priority level of the lifecycle phase (see `when`) that this observation will record.
 
     """
 
@@ -238,6 +248,7 @@ class StratifiedObservation(Observation):
         aggregator_sources: list[str] | None,
         aggregator: Callable[[pd.DataFrame], float | pd.Series[float]],
         to_observe: Callable[[Event], bool] = lambda event: True,
+        priority: int = DEFAULT_EVENT_PRIORITY,
     ):
         super().__init__(
             name=name,
@@ -249,6 +260,7 @@ class StratifiedObservation(Observation):
             results_updater=results_updater,
             results_formatter=results_formatter,
             to_observe=to_observe,
+            priority=priority,
         )
         self.aggregator_sources = aggregator_sources
         self.aggregator = aggregator
@@ -422,6 +434,8 @@ class AddingObservation(StratifiedObservation):
         Method or function that computes the quantity for this observation.
     to_observe
         Method or function that determines whether to perform an observation on this Event.
+    priority
+        The priority level of the lifecycle phase (see `when`) that this observation will record.
 
     """
 
@@ -435,6 +449,7 @@ class AddingObservation(StratifiedObservation):
         aggregator_sources: list[str] | None,
         aggregator: Callable[[pd.DataFrame], float | pd.Series[float]],
         to_observe: Callable[[Event], bool] = lambda event: True,
+        priority: int = DEFAULT_EVENT_PRIORITY,
     ):
         super().__init__(
             name=name,
@@ -446,6 +461,7 @@ class AddingObservation(StratifiedObservation):
             aggregator_sources=aggregator_sources,
             aggregator=aggregator,
             to_observe=to_observe,
+            priority=priority,
         )
 
     @staticmethod
@@ -507,6 +523,8 @@ class ConcatenatingObservation(UnstratifiedObservation):
         Method or function that formats the raw observation results.
     to_observe
         Method or function that determines whether to perform an observation on this Event.
+    priority
+        The priority level of the lifecycle phase (see `when`) that this observation will record.
 
     """
 
@@ -518,6 +536,7 @@ class ConcatenatingObservation(UnstratifiedObservation):
         requires_attributes: list[str],
         results_formatter: Callable[[str, pd.DataFrame], pd.DataFrame],
         to_observe: Callable[[Event], bool] = lambda event: True,
+        priority: int = DEFAULT_EVENT_PRIORITY,
     ):
         requires_attributes = ["event_time"] + requires_attributes
         super().__init__(
@@ -529,6 +548,7 @@ class ConcatenatingObservation(UnstratifiedObservation):
             results_updater=self.concatenate_results,
             results_formatter=results_formatter,
             to_observe=to_observe,
+            priority=priority,
         )
 
     def get_results_of_interest(self, pop: pd.DataFrame) -> pd.DataFrame: